feat(typescript): Add ts to Project
