--- .eslintrc.json
@@ -1,36 +1,0 @@
-{
-    "extends": [
-        "eslint:recommended",
-        "plugin:import/errors",
-        "plugin:react/recommended",
-        "plugin:jsx-a11y/recommended",
-        "plugin:prettier/recommended"  
-    ],
-    "rules":{
-        "react/prop-types": 0,
-        "no-console": 1
-    },
-    "plugins": [
-        "react",
-        "import",
-        "jsx-a11y"
-    ],
-
-    "parser": "babel-eslint",
-
-    "parserOptions": {
-        "ecmaVersion": 2018,
-        "sourceType": "module",
-        "ecmaFeatures": {
-            "jsx": true
-        }
-    },
-
-    "env": {
-        "es6": true,
-        "browser": true,
-        "node": true, 
-        "jest": true
-    }
-     
-}
