+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "format:base": "npm run prettier -- \"{src,e2e}/**/*{.js,.json,.css,.scss}\" ",
-    "lint": "eslint \"src/ **/*.{js, jsx}\" ",
+    "lint": "tslint --project .",
     "test": "jest --silent",
     "test:update": "jest --silent -u",
     "test:watch": "jest --silent --watch",
@@ -19,22 +19,24 @@
     "@babel/core": "^7.4.5",
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
+    "@types/reach__router": "^1.2.4",
+    "@types/react": "^16.8.19",
+    "@types/react-dom": "^16.8.4",
     "babel-core": "^6.26.3",
-    "babel-eslint": "^10.0.1",
     "babel-loader": "^7.1.5",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "eslint": "^5.16.0",
-    "eslint-config-prettier": "^4.2.0",
-    "eslint-plugin-import": "^2.17.2",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-prettier": "^3.0.1",
-    "eslint-plugin-react": "^7.13.0",
     "jest": "^24.8.0",
     "parcel-bundler": "^1.12.3",
     "prettier": "^1.17.0",
-    "react-test-renderer": "^16.8.6"
+    "react-test-renderer": "^16.8.6",
+    "tslint": "^5.16.0",
+    "tslint-config-prettier": "^1.18.0",
+    "tslint-react": "^4.0.0",
+    "typescript": "^3.4.5"
   },
   "dependencies": {
     "@emotion/core": "^10.0.10",
