+++ tslint.json
@@ -1,0 +1,15 @@
+{
+    "extends": [
+        "tslint:recommended",
+        "tslint-react",
+        "tslint-config-prettier"
+    ],
+    "rules": {
+        "ordered-imports": false,
+        "object-literal-sort-keys": false,
+        "interface-name": false,
+        "member-ordering": false,
+        "no-console": false
+
+    }
+}
